Move `hashbrown` and `foldhash` out of `bevy_utils`
