+++ crates/bevy_platform_support/Cargo.toml
@@ -11,6 +11,11 @@ keywords = ["bevy"]
 [features]
 default = ["std"]
 
+# Functionality
+
+## Adds serialization support through `serde`.
+serde = ["hashbrown/serde"]
+
 # Platform Compatibility
 
 ## Allows access to the `std` crate. Enabling this feature will prevent compilation
@@ -21,9 +26,10 @@ std = [
   "critical-section?/std",
   "portable-atomic?/std",
   "portable-atomic-util?/std",
+  "foldhash/std",
 ]
 
-alloc = ["portable-atomic-util?/alloc"]
+alloc = ["portable-atomic-util?/alloc", "dep:hashbrown"]
 
 ## `critical-section` provides the building blocks for synchronization primitives
 ## on all platforms, including `no_std`.
@@ -39,6 +45,11 @@ portable-atomic = { version = "1", default-features = false, features = [
   "fallback",
 ], optional = true }
 portable-atomic-util = { version = "0.2.4", default-features = false, optional = true }
+foldhash = { version = "0.1.3", default-features = false }
+hashbrown = { version = "0.15.1", features = [
+  "equivalent",
+  "raw-entry",
+], optional = true, default-features = false }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 web-time = { version = "1.1", default-features = false }
