+++ crates/bevy_utils/Cargo.toml
@@ -14,17 +14,17 @@ default = ["std", "serde"]
 # Functionality
 
 ## Adds serialization support through `serde`.
-serde = ["hashbrown/serde"]
+serde = ["bevy_platform_support/serde"]
 
 # Platform Compatibility
 
 ## Allows access to the `std` crate. Enabling this feature will prevent compilation
 ## on `no_std` targets, but provides access to certain additional features on
 ## supported platforms.
-std = ["alloc", "bevy_platform_support/std", "foldhash/std", "dep:thread_local"]
+std = ["alloc", "bevy_platform_support/std", "dep:thread_local"]
 
 ## Allows access to the `alloc` crate.
-alloc = ["bevy_platform_support/alloc", "hashbrown"]
+alloc = ["bevy_platform_support/alloc"]
 
 ## `critical-section` provides the building blocks for synchronization primitives
 ## on all platforms, including `no_std`.
@@ -37,11 +37,6 @@ portable-atomic = ["bevy_platform_support/portable-atomic"]
 [dependencies]
 bevy_platform_support = { path = "../bevy_platform_support", version = "0.16.0-dev", default-features = false }
 
-foldhash = { version = "0.1.3", default-features = false }
-hashbrown = { version = "0.15.1", features = [
-  "equivalent",
-  "raw-entry",
-], optional = true, default-features = false }
 thread_local = { version = "1.0", optional = true }
 
 [dev-dependencies]
